start implementation of units-for-slot
